+++ Mark/week2/app/package.json
@@ -21,7 +21,8 @@
     "react-native-safe-area-context": "^4.10.3",
     "react-native-screens": "^3.31.1",
     "react-native-svg": "^15.3.0",
-    "react-native-svg-transformer": "^1.4.0"
+    "react-native-svg-transformer": "^1.4.0",
+    "react-native-webview": "^13.10.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
